fixed build issues
building now properly includes the icons and styles. why didn't it do
that in the first place? idk. is it possible that it's i tried to use
bun and pnpm at the same time thinking that it would have no
consequences? maybe. but everything is now well and good (for now).